Refactor helm.go into cli package, split commands up across multiple files.
